Split brain_stdlib into multiple modules and finally kill it.
